Creating Analytics App
This commit introduces a brand-new analytics app.
Its purpose is to track relevant analytics about Shelf Help
for both performance of the application and curiosity in how
it is being used.

This involves a new frontend app that is a simple single page
with client-side queries to the database. With this we can
inspect, visualize, and analyze behavior of Shelf Help.