Frontend Rewrite
This rewrite cleans up a lot of unused packages along with
updating many of the existing libraries. This was mainly
to see if we could get the status bar to change colors in iOS,
and we discovered that the PWA and/or its manifest is the limiting
factor.

If you comment out the PWA manifest, then the status bar will
switch with the dark/light mode switch in Safari. However,
the PWA is more valuable to keep. So this ends up cleaning
things up while establishing that baseline knowledge for
future research.