Adding improved loading UI to Search mode
Makes the Search Multiselect have a blurred overlay instead of flashing.
Tried having it load as well like the spinner, but having two spinners
(one on the button, and one on the search bar, each a different color)
did not look right. I think this is less distracting and improves the look/feel.