Fixing Status Bar color in PWA
This change involved a lot of trial and error.
Ultimately, what made the behavior work best was adding the
theme-color meta tag and updating the callbacks to update that tag.
One minor detail is that the PWA initially uses whatever the tag is set at
in the index.html, so a useEffect was required to find whether we were
in a dark/light state, when opening the PWA. That allows a good experience
when closing and opening the app and maintaining its state.