Add support for multiple libraries with same libby id
Some libraries share a Libby library id due to being part of an
e-book consortium. This introduces a new pattern that allows use to
have a single libby id to share multiple libraries. We will still use
the libby library id as the main key, but we can add net-new ones,
and then just add them to the libraryToLibbyId dictionary/map.

This commit also adds more libraries to be rendered in the multi-select
so that it appears more searchable.